pyproject.toml
[project]
name = "bookified-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.3",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "reportlab>=4.4.10",
    "ruff>=0.15.7",
    "ty>=0.0.25",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
exclude = ["src/db/database_types.py", "src/db/schema.sql"]

[tool.ruff.format]
exclude = ["src/db/database_types.py", "src/db/schema.sql"]

[tool.ruff.lint]
select = ["I"]